apps/tinyolly/tests/test_compression.py
# BSD 3-Clause License
#
# Copyright (c) 2025, Infrastructure Architects, LLC
# All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions are met:
#
# 1. Redistributions of source code must retain the above copyright notice,
#    this list of conditions and the following disclaimer.
#
# 2. Redistributions in binary form must reproduce the above copyright notice,
#    this list of conditions and the following disclaimer in the documentation
#    and/or other materials provided with the distribution.
#
# 3. Neither the name of the copyright holder nor the names of its
#    contributors may be used to endorse or promote products derived from this
#    software without specific prior written permission.
#
# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

"""
Tests for data compression.

These tests verify:
- ZSTD compression of large payloads
- Decompression on retrieval
- Compression threshold behavior
"""
import pytest


class TestCompression:
    """Tests for data compression."""

    @pytest.mark.asyncio
    async def test_compress_large_span_attributes(self, storage):
        """Test compression of large span attributes."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_compress_large_log_body(self, storage):
        """Test compression of large log bodies."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_decompress_on_read(self, storage):
        """Test automatic decompression when reading."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_small_data_not_compressed(self, storage):
        """Test that small data is not compressed."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_compression_threshold(self, storage):
        """Test compression threshold configuration."""
        pytest.skip("SQLite storage not yet implemented")

    @pytest.mark.asyncio
    async def test_binary_data_compression(self, storage):
        """Test compression of binary attribute values."""
        pytest.skip("SQLite storage not yet implemented")
