apps/tinyolly/requirements.txt
# TinyOlly Unified Requirements
# Combines UI and OTLP Receiver dependencies

# Core FastAPI/Web (UI mode)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
uvloop>=0.19.0
psutil>=5.9.0
aiohttp>=3.9.0
pyyaml>=6.0

# gRPC (Receiver mode)
grpcio>=1.60.0
protobuf>=5.0.0
opentelemetry-proto>=1.24.0

# OpenTelemetry (Both modes)
opentelemetry-api>=1.22.0
opentelemetry-distro>=0.42b0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-logging>=0.42b0

# Storage (from common module)
redis>=4.5.0,<6.0.0
zstandard>=0.21.0,<1.0.0
msgpack>=1.0.0,<2.0.0
orjson>=3.9.0,<4.0.0
async-lru>=2.0.0,<3.0.0
